Implement pre-reboot hook.
The pre-reboot hook is a subset of the boostrap functionality; it
doesn't trigger an OS update, and it doesn't touch the kubernetes
configuration. It just updates the torcx packages.

It can mark completion by setting a node annotation; this is to be used
with the Container-Linux-Update-Operator hook feature.

On completion, it will remove any torcx stores  that are known to be
superfluous. This means versioned directories for versions that do not
exist on the system.

This commit also removes the post-reboot hook, which is not needed.