hook: gather k8s version from local state only
This changes hooks behavior when determining kubelet version.
Before, it was always querying the api-server first and falling back
to kubelet.env. Now, it only uses local state from the env file.

This is to avoid a race between KVO and CLUO, where KVO has not yet
scheduled a node for kubelet upgrade but tectonic-torcx is already
acting on the new version due to api-server detection.