kube: hard-fail on client building failures
When gathering cluster version from remote api-servers, failing to
build the client is now treated as an hard failure which should not be
ignored, as it is not transient and likely due to an underlying problem
in the environment setup.

This was recently discovered in FR-497 as triggered by a race in the
way the bare-metal installer provision `kubeconfig`.
While this does not fix the issue, it makes the problem easier to
notice and allows for graceful retries.